pyproject.toml
[project]
name = "stock-purchase-analyzer"
version = "0.1.0"
description = "Multi-agent AI stock analysis system using social media sentiment, fundamentals, and historical patterns"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    # Data & Finance
    "yfinance>=0.2.40",
    "pandas>=2.2.0",
    "pandas-ta>=0.3.14b1",
    "numpy>=1.26.0",
    # AI/LLM
    "google-generativeai>=0.8.0",
    # Database
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    # Settings & Validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    # UI
    "streamlit>=1.40.0",
    # Caching
    "cachetools>=5.5.0",
    # HTTP
    "httpx>=0.27.0",
    # Environment
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
]
sentiment = [
    "praw>=7.7.0",
    "newsapi-python>=0.2.7",
    "finvizfinance>=0.14.0",
]
economic = [
    "fredapi>=0.5.2",
]
ollama = [
    "ollama>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "config"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
